+++ ReleaseNotes.md
@@ -1,4 +1,10 @@
-# v0.1.1
-Updater test
+# v0.1.2
 
+## Main Page
+1. Added an hour input selection
+
+### Settings Page
+1. Adding the settings page! Which currently just has the version number :)
+
+### Misc
 1. Updated icons